+++ src/model/person/Relation.java
@@ -1,0 +1,7 @@
+package model.person;
+
+import java.io.Serializable;
+
+public enum Relation implements Serializable {
+    Person, Spouse, Mother, Father, Sister, Brother, Son, Daughter
+}
