+++ src/model/person/Sex.java
@@ -1,0 +1,7 @@
+package model.person;
+
+import java.io.Serializable;
+
+public enum Sex implements Serializable {
+    Male, Female
+}
