--- src/person/Relation.java
@@ -1,7 +1,0 @@
-package person;
-
-import java.io.Serializable;
-
-public enum Relation implements Serializable {
-    Person, Spouse, Mother, Father, Sister, Brother, Son, Daughter
-}
