--- src/person/Sex.java
@@ -1,7 +1,0 @@
-package person;
-
-import java.io.Serializable;
-
-public enum Sex implements Serializable {
-    Male, Female
-}
