+++ src/view/View.java
@@ -1,0 +1,9 @@
+package view;
+
+import presenter.Presenter;
+
+public interface View {
+    void start();
+    void setPresenter(Presenter presenter);
+    void print(String text);
+}
